feat: Bump observablehq runtime support
