+++ packages/observablehq-compiler/src/index.ts
@@ -1,7 +1,6 @@
 export type { ohq } from "./observable-shim.ts";
 
 export * from "./compiler.ts";
-export { ojs2notebook, omd2notebook, omd2notebookKit, ojs2notebookKit, download, html2notebook, notebook2html } from "./util.ts";
+export { ojs2notebook, omd2notebook, omd2notebookKit, ojs2notebookKit, download } from "./util.ts";
+export { compile as compileKit, html2notebook, notebook2html } from "./kit/index.ts";
 export * from "./writer.ts";
-
-import "../src/index.css";
