+++ packages/observablehq-compiler/tsconfig.json
@@ -20,6 +20,7 @@
         ]
     },
     "include": [
-        "./src/index.ts"
+        "./src/index.ts",
+        "src/kit/runtime.ts"
     ]
 }
